+++ src/extension.ts
@@ -43,6 +43,9 @@ export function activate(context: vscode.ExtensionContext) {
             `${extensionName}.stop`,
             () => extensionInstance.stopProcess(),
         ),
+        vscode.commands.registerTextEditorCommand('io.orta.jest.show-channel', ()=> {
+            channel.show();
+        }),
         ...registerFileChangeWatchers(extensionInstance),
     );
 }