+++ src/IPluginSettings.ts
@@ -1,5 +1,6 @@
 export interface IPluginSettings {
     autoEnable?: boolean;
+    enableInlineErrorMessages?: boolean;
     pathToJest?: string;
     pathToConfig?: string;
     rootPath?: string;