+++ src/IPluginSettings.ts
@@ -1,6 +1,7 @@
 export interface IPluginSettings {
     autoEnable?: boolean;
     enableInlineErrorMessages?: boolean;
+    enableSnapshotUpdateMessages?: boolean;
     pathToJest?: string;
     pathToConfig?: string;
     rootPath?: string;