+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Master
+
+* Show channel command - orta
+
 ### 1.6.4
 
 * Windows + Create React App improvements - brentatkins